@@ -1,11 +1,10 @@
-'''Insert header Here'''
-
-
-from pathlib import Path
-from .sqldbwrpr import *
-
-
-_PROJ_DESC = __doc__.split('\n')[0]
-_PROJ_PATH = Path(__file__)
-_PROJ_NAME = _PROJ_PATH.stem
-_PROJ_VERSION = '0.0.1'
+# '''Insert header Here'''
+# from pathlib import Path
+#
+# from .sqldbwrpr import *
+#
+#
+# _PROJ_DESC = __doc__.split('\n')[0]
+# _PROJ_PATH = Path(__file__)
+# _PROJ_NAME = _PROJ_PATH.stem
+# _PROJ_VERSION = '0.0.1'
